+++ src/config/general.rs
@@ -75,9 +75,6 @@ pub struct General {
     #[serde(default = "General::default_retain_connections_time")]
     pub retain_connections_time: u64,
 
-    #[serde(default = "General::default_server_round_robin")] // False
-    pub server_round_robin: bool,
-
     #[serde(default = "General::default_sync_server_parameters")] // False
     pub sync_server_parameters: bool,
 
@@ -251,10 +248,6 @@ impl General {
         4
     }
 
-    pub fn default_server_round_robin() -> bool {
-        false
-    }
-
     pub fn default_prepared_statements_cache_size() -> usize {
         8 * 1024
     }
@@ -350,7 +343,6 @@ impl Default for General {
             admin_password: String::from("admin"),
             server_lifetime: Self::default_server_lifetime(),
             retain_connections_time: Self::default_retain_connections_time(),
-            server_round_robin: Self::default_server_round_robin(),
             prepared_statements: Self::default_prepared_statements(),
             prepared_statements_cache_size: Self::default_prepared_statements_cache_size(),
             hba: Self::default_hba(),